11-06-2025 Charan-anil merged code

+++ DHA/src/main/java/com/ac/dha/repository/UploadERxAuthorizationRepository.java
@@ -2,10 +2,10 @@ package com.ac.dha.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.ac.dha.entties.UploadERxAuthorization;
+import com.ac.dha.entities.UploadERxAuthorization;
+
 @Repository
-public interface UploadERxAuthorizationRepository extends JpaRepository<UploadERxAuthorization, Integer>{
+public interface UploadERxAuthorizationRepository extends JpaRepository<UploadERxAuthorization, Long> {
 
 }
